EclTimesteppingParams: introduce translation unit
